changed enum def, refactored auth types, removed decorators
