+++ src/server.ts
@@ -6,14 +6,15 @@ import cors from 'cors';
 import express from 'express';
 import yaml from 'yamljs';
 import * as swaggerUi from 'swagger-ui-express';
+import {authorize} from "./middlewares/autorization/auth-util.js";
 
 const app = express();
 
 function setupExpress() {
 	app.use(cors({ origin: true }));
 	app.route('/').get(router.root);
-	app.route(config.requestRoute).get(router.getIdForEntity);
-	app.route(config.requestRoute + '/find').get(router.findIdForEntity);
+	app.route(config.requestRoute).get(authorize('CREATE'), router.getIdForEntity);
+	app.route(config.requestRoute + '/find').get(authorize('FIND'), router.findIdForEntity);
 	app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(yaml.load('src/resources/swagger.yaml')));
 	app.use(defaultErrorHandler);
 }
